ci: only build master branch builds
By default Travis Ci will run a build for the current `HEAD` of the
branch the PR is on as well as the state of `master` after merging the
PR branch. This changes it to only run on the `master` branch, with the
exclusion of GreenKeeper's special builds.

The net effect of this is PRs will only have one build to wait on.